requirements.txt
Flask
python-docx
Pillow
docx2pdf
pywin32==311; sys_platform == "win32"
gunicorn
psycopg2-binary
Flask-SQLAlchemy
reportlab
python-dotenv